requirements.txt
groq
pyyaml
